Fix issue #1073: Move the SpatialFactory and related to Microsoft.Spatial library
