+++ src/Microsoft.Spatial/Microsoft.Spatial.csproj
@@ -43,6 +43,8 @@
     <Compile Include="CompositeKey.cs" />
     <Compile Include="FormatterExtensions.cs" />
     <Compile Include="GeographyPipeline.cs" />
+    <Compile Include="GeographyFactory.cs" />
+    <Compile Include="GeographyFactoryOfT.cs" />
     <Compile Include="GeographyFullGlobe.cs" />
     <Compile Include="GeographyCurve.cs" />
     <Compile Include="GeographyLineString.cs" />
@@ -59,6 +61,8 @@
     <Compile Include="GeometryPipeline.cs" />
     <Compile Include="GeometryCollection.cs" />
     <Compile Include="GeometryCurve.cs" />
+    <Compile Include="GeometryFactory.cs" />
+    <Compile Include="GeometryFactoryOfT.cs" />
     <Compile Include="GeometryLineString.cs" />
     <Compile Include="GeometryMultiCurve.cs" />
     <Compile Include="GeometryMultiLineString.cs" />
@@ -76,6 +80,7 @@
     <Compile Include="GeoJsonObjectFormatter.cs" />
     <Compile Include="ParseErrorException.cs" />
     <Compile Include="SpatialBuilder.cs" />
+    <Compile Include="SpatialFactory.cs" />
     <Compile Include="SpatialFormatter.cs" />
     <Compile Include="SpatialReader.cs" />
     <Compile Include="SpatialType.cs" />
