+++ test/FunctionalTests/Microsoft.OData.TestCommon/Build.NetFramework/Microsoft.OData.TestCommon.csproj
@@ -44,7 +44,6 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="..\Properties\AssemblyInfo.cs" />
-    <Compile Include="..\SpatialFactory.cs" />
   </ItemGroup>
   <ItemGroup>
     <Service Include="{82A7F48D-3B50-4B1E-B82E-3ADA8210C358}" />
