+++ test/FunctionalTests/Tests/DataOData/Common/OData/Microsoft.Test.Taupo.OData.csproj
@@ -67,9 +67,6 @@
   <!-- Source files -->
   <ItemGroup>
     <Compile Include="$(DataFxIncPath)\AssemblyRefs.cs" />
-    <Compile Include="$(EnlistmentRoot)\test\FunctionalTests\Microsoft.OData.TestCommon\SpatialFactory.cs">
-      <Link>Common\SpatialFactory.cs</Link>
-    </Compile>
     <Compile Include="Atom\TestAtomConstants.cs" />
     <Compile Include="Atom\AtomMetadataBuilder.cs" />
     <Compile Include="Atom\ODataPayloadElementAtomMetadataNormalizer.cs" />
